requirements.txt
tornado==6.0.3
python-telegram-bot==12.1.1
attrs==19.1.0
tornado-sqlalchemy
psycopg2
vk_api